Fix: Complete color definitions, Camera2 API, ProControlsPanel params, and permissions
- Add all missing color values (GlassWhite, GlassBlack, AccentPurple, AccentPink, ErrorRed, etc)
- Fix Camera2 Control API: setCaptureRequestOptions instead of applyCaptureRequestOptions
- Fix ProControlsPanel parameter names to match function signature
- Fix permission shouldShowRationale access
- Fix type mismatch in ISO slider (Int to Float conversion)

+++ app/src/main/kotlin/com/retrocam/app/ui/components/ProControlsPanel.kt
@@ -74,7 +74,7 @@ fun ProControlsPanel(
             capabilities?.isoRange?.let { range ->
                 ManualControlSlider(
                     label = "ISO",
-                    value = currentSettings.iso ?: range.first,
+                    value = (currentSettings.iso ?: range.first).toFloat(),
                     valueRange = range.first.toFloat()..range.last.toFloat(),
                     onValueChange = { onSettingsChange(currentSettings.copy(iso = it.roundToInt())) },
                     valueFormatter = { it.roundToInt().toString() },
